+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADEntitlementManagementConnectedOrganization/MSFT_AADEntitlementManagementConnectedOrganization.psm1
@@ -89,20 +89,20 @@ function Get-TargetResource
 
         $getValue = $null
 
-        if (-not [System.String]::IsNullOrEmpty($id))
+        if (-not [System.String]::IsNullOrEmpty($Id))
         {
-            $getValue = Get-MgBetaEntitlementManagementConnectedOrganization -ConnectedOrganizationId $id `
+            $getValue = Get-MgBetaEntitlementManagementConnectedOrganization -ConnectedOrganizationId $Id `
                 -ErrorAction SilentlyContinue
         }
 
         if ($null -eq $getValue)
         {
-            if (-not [System.String]::IsNullOrEmpty($id))
+            if (-not [System.String]::IsNullOrEmpty($Id))
             {
-                Write-Verbose -Message "Entitlement Management Connected Organization with id {$id} was not found."
+                Write-Verbose -Message "Entitlement Management Connected Organization with id {$Id} was not found."
             }
 
-            if (-Not [string]::IsNullOrEmpty($DisplayName))
+            if (-not [string]::IsNullOrEmpty($DisplayName))
             {
                 $getValue = Get-MgBetaEntitlementManagementConnectedOrganization `
                     -Filter "displayName eq '$DisplayName'" `
@@ -787,7 +787,7 @@ function Export-TargetResource
             Write-M365DSCHost -Message "    |---[$i/$($getValue.Count)] $displayedKey" -DeferWrite
             $params = @{
                 id                    = $config.id
-                DisplayName           = $config.displayName
+                DisplayName           = $displayedKey
                 Ensure                = 'Present'
                 Credential            = $Credential
                 ApplicationId         = $ApplicationId