+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -2347,9 +2347,19 @@ function Get-IntuneSettingCatalogPolicySettingDSCValue
 
     $key = Get-SettingsCatalogSettingName -SettingDefinition $SettingDefinition -AllSettingDefinitions $AllSettingDefinitions
 
-    if (-not $DSCParams.ContainsKey($key))
+    if ($DSCParams.Keys -notcontains $key)
     {
         return $null
+    }
+
+    # Fixes potential case sensitivity issue.
+    foreach ($hashKey in $DSCParams.Keys)
+    {
+        if ($hashKey -eq $key)
+        {
+            $key = $hashKey
+            break
+        }
     }
 
     $isArray = $false
@@ -2701,7 +2711,7 @@ function Update-IntuneDeviceConfigurationPolicy
         [Array]
         $Settings
     )
-
+    $VerbosePreference = 'continue'
     try
     {
         $Uri = (Get-MSCloudLoginConnectionProfile -Workload MicrosoftGraph).ResourceUrl + "beta/deviceManagement/configurationPolicies/$DeviceConfigurationPolicyId"
@@ -2715,7 +2725,7 @@ function Update-IntuneDeviceConfigurationPolicy
             'settings'          = $Settings
         }
         $body = $policy | ConvertTo-Json -Depth 20
-        # Write-Verbose -Message $body -Verbose
+        Write-Verbose -Message "Updating policy with:`r`n$body"
         Invoke-MgGraphRequest -Method PUT -Uri $Uri -Body $body -ErrorAction Stop
     }
     catch