+++ CHANGELOG.md
@@ -2,12 +2,26 @@
 
 # UNRELEASED
 
+* AADEntitlementManagementConnectedOrganization
+  * Fixed an issue where an empty display name would result in a parameter binding exception.
+    FIXES [#4981](https://github.com/microsoft/Microsoft365DSC/issues/4981)
 * EXOHostedContentFilterPolicy
   * Fixed an issue with setting the default value for `IntraOrgFilterState`.
     FIXES [#5031](https://github.com/microsoft/Microsoft365DSC/issues/5031)
+* IntuneAppProtectionPolicyiOS
+  * Fixed empty array comparison.
+    FIXES [#5736](https://github.com/microsoft/Microsoft365DSC/issues/5736)
+* IntuneWifiConfigurationPolicyAndroidEnterpriseWorkProfile
+  * Fixes the data type for extraction and creation.
+* SPOSearchManagedProperty
+  * Fixed an issue where the connection url was not the admin url required by the resource.
+    FIXES [#5093](https://github.com/microsoft/Microsoft365DSC/issues/5093)
+    FIXES [#4538](https://github.com/microsoft/Microsoft365DSC/issues/4538)
 * DEPENDENCIES
   * Updated Microsoft.Graph to version 2.27.0.
   * Updated MSCoudLoginAssistant to version 1.1.43.
+* MISC
+  * Fixes case sensitivity comparison when updating device configuration policies.
 
 # 1.25.416.1